+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - **Standalone** Signing OSX app bundle
 - **Plugin** Fix Apple Logic auval crashes
 - **Plugin** Raise audio channels/calls from 4 to 6
+- **Plugin** Add option auto-mix-n-1
 - **Misc** Fix possible Ghost Calls (rewrite call id handling)
 - **Misc** Support OSX Versions <10.11
 - **Misc** Improve vumeter warning state